% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{boxPlotList}
\alias{boxPlotList}
\title{New Boxplot Function}
\usage{
boxPlotList(
  dat,
  l.cells = NULL,
  dat.name = "c.dat",
  col.name = NULL,
  jitter.f = 0.5,
  pts = T,
  notchs = F,
  bplog = "y",
  sort = T
)
}
\arguments{
\item{dat}{: Experiment list RD.}

\item{col.name}{c("area","mean.gfp.start","mean.cy5.start")}

\item{l.cells:}{Cells in a list format}

\item{dat.name:}{Dataframe to pull data from}

\item{col.name:}{collumn name  to get data for the boxplot}

\item{jitter.f:}{Factor of jitter to  accomplish}

\item{pts:}{points to add to boxplot}

\item{notchs:}{logical (T/F) stand for notch selection}
}
\description{
New Boxplot Function
}
