% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{bp.selector.advanced}
\alias{bp.selector.advanced}
\title{Interactive stat maker, min max normalization.}
\usage{
bp.selector.advanced(
  dat,
  cell = NULL,
  cells = NULL,
  dat.name = NULL,
  plot.new = T,
  save.bp = F,
  view.cells = F,
  env = NULL,
  localize = T
)
}
\description{
Interactive stat maker, min max normalization.
}
