% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{bp.selector}
\alias{bp.selector}
\title{Interactive statistic maker. This creates a statistic based on peak heights or peak areas.}
\usage{
bp.selector(
  dat,
  cell = NULL,
  cells = NULL,
  groups = NULL,
  dat.name = NULL,
  plot.new = T,
  save.bp = F,
  view.cells = F,
  env = NULL,
  statType = "minMax"
)
}
\description{
Interactive statistic maker. This creates a statistic based on peak heights or peak areas.
}
