% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{bScore}
\alias{bScore}
\title{binary score for all cells for the regions of interest bScore
returns the scoring for all cells subject to the above parameters.
as well as the sum for the snr scores and the sd for the snr scores.}
\usage{
bScore(blc, snr, snr.lim, blc.lim, levs, wr, cnames = NULL)
}
\arguments{
\item{blc}{is the baseline corrected data}

\item{snr}{is the signal to noise peak data}

\item{snr.lim}{is the threshold for significance on the peaks}

\item{blc.lim}{is the intensity above baseline theshold}

\item{levs}{indicates the regions of interest. (e.g. the response windows for which the cells will be scored)}

\item{wr}{indicates the response windows.}

\item{cnames}{indicates the cells to score (if null all cells will be scored)}
}
\description{
binary score for all cells for the regions of interest bScore
returns the scoring for all cells subject to the above parameters.
as well as the sum for the snr scores and the sd for the snr scores.
}
