% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{bscore2}
\alias{bscore2}
\title{Binary scoring dependent upon score const pharm talbe values
 Best way to determine parameters is to look through trace click before hand}
\usage{
bscore2(dat, levs.1 = NULL, snr.lim = 5, blc.lim = 0.05)
}
\arguments{
\item{snr.min}{= minimun signal to noise value}

\item{max.min=}{minimun above baseline threshold}

\item{tot.min=}{area minimun to consider}

\item{wm.min=}{which max, Where within the window region does the maximun value occur}

\item{wm.max=}{where to stop looking for the maximun value}
}
\description{
Binary scoring dependent upon score const pharm talbe values
 Best way to determine parameters is to look through trace click before hand
}
