% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{c.sort}
\alias{c.sort}
\title{function to obtained sorted cell names based off 
collumn names from c.dat and bin}
\usage{
\method{c}{sort}(dat, char = NULL)
}
\description{
function to obtained sorted cell names based off 
collumn names from c.dat and bin
}
