% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{cell.view}
\alias{cell.view}
\title{Function to View all cells in grid, zoomed in.}
\usage{
cell.view(
  dat,
  cell = NULL,
  img = NULL,
  zoom = TRUE,
  cols = NULL,
  lmain = "",
  bcex = 0.8,
  labs = T,
  plot.new = T,
  cell.name = T
)
}
\description{
Function to View all cells in grid, zoomed in.
}
