% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{cell_type_modeler}
\alias{cell_type_modeler}
\title{Using marios cell type models I will create the cell_type model list. This is BETA version,
Simply inputing a RD.experiment which contains AMCK, R3J, and and image of IB4 only at
img3 and an image of GFP only at img4, this function returns a list of model output matrices.}
\usage{
cell_type_modeler(dat, modelDir = NA)
}
\description{
Using marios cell type models I will create the cell_type model list. This is BETA version,
Simply inputing a RD.experiment which contains AMCK, R3J, and and image of IB4 only at
img3 and an image of GFP only at img4, this function returns a list of model output matrices.
}
