% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTyping.R
\name{Cell_Typer_2}
\alias{Cell_Typer_2}
\title{Function with 3 options_ Edit_ct, classify UL , classify thermos
This follows Marios scheme for classifying our cell types
edit_ct=Logical, if true each cell class will be double checked
UL_classify= If TRUE then classify large diameter cells
GFP=logical, if TRUE then classify green cells
cell_types=list input.  This is mainly used if the large cell types have already been classified.
if so then then the large cell types are passed straight to the cell_types
181016 If the large diameter cells have been classified, then do not score again.}
\usage{
Cell_Typer_2(tmp_rd, edit_ct = F, UL_classify = T, GFP = T, cell_types = NULL)
}
\description{
Function with 3 options_ Edit_ct, classify UL , classify thermos
This follows Marios scheme for classifying our cell types
edit_ct=Logical, if true each cell class will be double checked
UL_classify= If TRUE then classify large diameter cells
GFP=logical, if TRUE then classify green cells
cell_types=list input.  This is mainly used if the large cell types have already been classified.
if so then then the large cell types are passed straight to the cell_types
181016 If the large diameter cells have been classified, then do not score again.
}
