% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{cellTypeAdder}
\alias{cellTypeAdder}
\title{Add the original cell types to the RD.experiment
This fucntion will look for double classified cells and return 
what is double classified}
\usage{
cellTypeAdder(dat)
}
\description{
Add the original cell types to the RD.experiment
This fucntion will look for double classified cells and return 
what is double classified
}
