% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{cellzand}
\alias{cellzand}
\title{Function to select rows based on collumn parameters
dat can be either a raw RD object or an RD dataframe
ex dat -or- dat$bin}
\usage{
cellzand(dat, collumn = NULL, parameter = 1, cells = NULL)
}
\description{
Function to select rows based on collumn parameters
dat can be either a raw RD object or an RD dataframe
ex dat -or- dat$bin
}
