% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTyping.R
\name{census.brewer}
\alias{census.brewer}
\title{This is the old way to create a census! Recently updated though}
\usage{
census.brewer(dat)
}
\description{
This is the old way to create a census! Recently updated though
}
