% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{census_to_table}
\alias{census_to_table}
\title{Legacy function to convert the census into binary collumns}
\usage{
census_to_table(dat)
}
\description{
Legacy function to convert the census into binary collumns
}
