% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{col_binner}
\alias{col_binner}
\title{Now lets create the function to put premade groups into a binary collumns}
\usage{
col_binner(dat, cells)
}
\description{
Now lets create the function to put premade groups into a binary collumns
}
