% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{CollectMulti}
\alias{CollectMulti}
\title{given a list of file names collect and merge all bin scp and c.dat data}
\usage{
CollectMulti(f.names, rd.names = NULL)
}
\description{
given a list of file names collect and merge all bin scp and c.dat data
}
