% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{combinner}
\alias{combinner}
\title{Create Binary Classes of cells
dat IS THE rd. INPUT
LEVS, IS AN OPTIONAL ARGUEMENT, IF LEFT BLACK THE FUNCTION WILL LOOK IN THE BIN DATA.FRAME COLLUMN ANMES
TO ALLOW TO MANUALLY SECT THE COLLUMNS YOU WANT TO COMBINE}
\usage{
combinner(dat, levs = NULL, bin_it = T)
}
\description{
Create Binary Classes of cells
dat IS THE rd. INPUT
LEVS, IS AN OPTIONAL ARGUEMENT, IF LEFT BLACK THE FUNCTION WILL LOOK IN THE BIN DATA.FRAME COLLUMN ANMES
TO ALLOW TO MANUALLY SECT THE COLLUMNS YOU WANT TO COMBINE
}
