% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{density_ct_plotter}
\alias{density_ct_plotter}
\title{This function plots the stat(data.frame form) of all cells, and individual
cell type densities.}
\usage{
density_ct_plotter(
  dat,
  cells,
  cell_types,
  stat = dat$c.dat["area"],
  xlim_top = NULL,
  xlim_bottom = NULL,
  overlay = T,
  dense_sep = T,
  plot_new = T,
  env = NULL,
  dat.name = NULL,
  abline_loc = 0
)
}
\arguments{
\item{dat}{RD data}

\item{cells}{Total group}

\item{cell_types}{How to seperate the groups}

\item{stat}{premade statistic in data.frame formate where row names are cell.names}

\item{xlim_top}{this is the maximun xlim value to display}

\item{xlim_bottom}{this is the minimun xlim value to display}

\item{overlay}{will plot the density plot ontop of the cells density plot}

\item{plot_new}{Will create a new window for this plot}

\item{abline_loc}{where to display the added line to help display data better}

\item{dens_sep}{This will plot out the densitys on seperate plots}
}
\description{
This function plots the stat(data.frame form) of all cells, and individual
cell type densities.
}
