% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{deStatMaker}
\alias{deStatMaker}
\title{Function to create a direct effect stat. 
this stat is automatically tested against the control to use.
Due to nature of these experiments there are ususally only one
window region really compare to, to observe the direct effect.
a major weakness of this statistic is taht direct effects cannot be automatically scored
if they are automatically score this would be significantly better.
since we are unable to compute this we compare window regions of anything. This confuses the outpu
of this analysis}
\usage{
deStatMaker(
  dat,
  statType = "tot",
  testPulseNames = "^[kK][.]30",
  controlToUse = 1
)
}
\description{
Function to create a direct effect stat. 
this stat is automatically tested against the control to use.
Due to nature of these experiments there are ususally only one
window region really compare to, to observe the direct effect.
a major weakness of this statistic is taht direct effects cannot be automatically scored
if they are automatically score this would be significantly better.
since we are unable to compute this we compare window regions of anything. This confuses the outpu
of this analysis
}
