% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ecdfPlotter}
\alias{ecdfPlotter}
\title{This function displayed the newly create stats in an automated way.
It depends on the cell types present to display the stat.
#' Adding more controls at the begining really improves the view of this}
\usage{
ecdfPlotter(
  dat,
  controlNames,
  testNames,
  legendSep = 0.2,
  rdName = NA,
  cell_types = NA
)
}
\arguments{
\item{dat}{RD.experiment}

\item{stat}{slection between 'de' which returns the direct effect stat or 'ide' which displays the indirect effect stat}

\item{controlToView}{a major highlighter of this vixualization is the ability to compare against a control window region}
}
\description{
This function displayed the newly create stats in an automated way.
It depends on the cell types present to display the stat.
#' Adding more controls at the begining really improves the view of this
}
