% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{fancyBin}
\alias{fancyBin}
\title{Function to use the neural networks within the python package}
\usage{
fancyBin(dat)
}
\arguments{
\item{dat}{is the RD file input}
}
\description{
Function to use the neural networks within the python package
}
