% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{imageExtractor}
\alias{imageExtractor}
\title{Function to convert the full image into individual cell images.}
\usage{
imageExtractor(dat, image = "img3", channel = 1, range = 20)
}
\arguments{
\item{dat}{is the RD.experiment to load in. You should have images loaded into this list}

\item{image}{is the character name of the image within your list}

\item{channel}{is the rgb channel of your image to extrace}

\item{range}{is the side in all directions from center to define your view}
}
\description{
Function to convert the full image into individual cell images.
}
