% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{ImageFillerv2}
\alias{ImageFillerv2}
\title{Add images to the RD.experiment file}
\usage{
ImageFillerv2(dat, img_name_vec)
}
\arguments{
\item{dat}{RD.experiment input}
}
\description{
Add images to the RD.experiment file
}
