% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{imageProbMaker}
\alias{imageProbMaker}
\title{Function to create the cy5.bin and the gfp.bin collumns in the binary
data frame. This also adds probability scores for the images}
\usage{
imageProbMaker(dat)
}
\description{
Function to create the cy5.bin and the gfp.bin collumns in the binary
data frame. This also adds probability scores for the images
}
