% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{kurtosis}
\alias{kurtosis}
\title{Function taken from e1071 package
This observes the tails of a distribution
big kurtosis means small tails: certain
small kurtosis means long tails: uncertain}
\usage{
kurtosis(x, na.rm = FALSE, type = 3)
}
\description{
Function taken from e1071 package
This observes the tails of a distribution
big kurtosis means small tails: certain
small kurtosis means long tails: uncertain
}
