% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceVisualization.R
\name{LinesEvery.5}
\alias{LinesEvery.5}
\title{LinesEvery same as .4 but has image at begining of trace and moves to pic plot at >10 
multipi does not work on this.  Instead, if greater than 10, the traces are plotted as a portrait orientation
Also window labels are rotated on axis and place on the bottom of the plot
I am also adding two more images to the left side of the plot
171009 added underline.  Helps to show irreversibility
171031 added dat.n for the name fotthe experiment}
\usage{
LinesEvery.5(
  dat,
  m.names,
  img = "img1",
  channel = NULL,
  pic.plot = TRUE,
  zf = NULL,
  t.type = "mp.1",
  snr = NULL,
  lmain = "",
  cols = "black",
  levs = NULL,
  levs.cols = "grey90",
  values = NULL,
  plot.new = T,
  sf = 1,
  lw = 1,
  bcex = 1,
  p.ht = 7,
  p.wd = 10,
  lns = T,
  pts = F,
  underline = T,
  dat.n = NULL
)
}
\description{
LinesEvery same as .4 but has image at begining of trace and moves to pic plot at >10 
multipi does not work on this.  Instead, if greater than 10, the traces are plotted as a portrait orientation
Also window labels are rotated on axis and place on the bottom of the plot
I am also adding two more images to the left side of the plot
171009 added underline.  Helps to show irreversibility
171031 added dat.n for the name fotthe experiment
}
