% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Function to track the various users data during analysis
The goal is to append a line to the log with}
\usage{
logger(functionName, timeInFunction, additionalInfo = NA)
}
\description{
Needs to happen before the function goes
functionName <- as.character(match.call())
timeInFunction <- (proc.time() - time1)[3]
}
