% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{modelChooser}
\alias{modelChooser}
\title{This function allows you to choose models to apply to your applications
This will allow you to choose variouse models and apply it to the applications you desire}
\usage{
modelChooser(dat, levPat = NA, modelName = NA, windowSize = 3)
}
\description{
This function allows you to choose models to apply to your applications
This will allow you to choose variouse models and apply it to the applications you desire
}
