% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{modelViewer}
\alias{modelViewer}
\title{Function to view the cell type models. Pretty good vis}
\usage{
modelViewer(dat, cell, plot.new = T)
}
\description{
Function to view the cell type models. Pretty good vis
}
