% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTyping.R
\name{named.list}
\alias{named.list}
\title{create a list that uses the names input for the names in the list}
\usage{
named.list(...)
}
\description{
create a list that uses the names input for the names in the list
}
