% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePrep.R
\name{NumBlanks}
\alias{NumBlanks}
\title{take a defined window vector and
number of the contiguos blank regions ("")}
\usage{
NumBlanks(x)
}
\description{
take a defined window vector and
number of the contiguos blank regions ("")
}
