% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceVisualization.R
\name{PeakFunc7}
\alias{PeakFunc7}
\title{This peak func allows for multiple t.types to be plotted
170515: added pts and lns: (logical)
added dat.n for insertation of the name for the rd file}
\usage{
PeakFunc7(
  dat,
  n.names,
  t.type = "t.dat",
  Plotit.trace = T,
  Plotit.both = F,
  info = T,
  lmain = NULL,
  bcex = 0.7,
  yvar = T,
  ylim.max = NULL,
  zf = 40,
  pts = T,
  lns = T,
  levs = NULL,
  underline = T,
  dat.n = ""
)
}
\description{
This peak func allows for multiple t.types to be plotted
170515: added pts and lns: (logical)
added dat.n for insertation of the name for the rd file
}
