% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{pf_summary}
\alias{pf_summary}
\title{This takes a pf and allows you to create a binarry table based on the barcode
Created in pf.function}
\usage{
pf_summary(dat, response_classes = NULL, pf_col = NULL)
}
\description{
This takes a pf and allows you to create a binarry table based on the barcode
Created in pf.function
}
