% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{pharming_harvest}
\alias{pharming_harvest}
\title{Main robust import function}
\usage{
pharming_harvest(
  main_dir = NULL,
  area_conversion = 1.625,
  img_name_vec = NULL,
  image_question = T
)
}
\arguments{
\item{area_conversion}{conversion factor from the object used from the experiment}

\item{img_name_vec}{vector of image names ex c('bf.gfp.cy5', 'bf.gfp')}

\item{image_question}{will ask you to fill in images if you don't have an img_name_vec}
}
\description{
Main robust import function
}
