% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{probMaker}
\alias{probMaker}
\title{Function to return probability scores from the models}
\usage{
probMaker(dat)
}
\arguments{
\item{dat}{is the RD file input}
}
\description{
Function to return probability scores from the models
}
