% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{ProcConstPharm}
\alias{ProcConstPharm}
\title{Function to perform the basline correction}
\usage{
ProcConstPharm(dat, shws = 2, phws = 20, bl.meth = "SNIP")
}
\description{
Function to perform the basline correction
}
