% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDView.R
\name{RDView_2}
\alias{RDView_2}
\title{RDVIew but for traces, also identifies the score of the neurons}
\usage{
RDView_2(dat, cells = NULL, levs = NULL)
}
\arguments{
\item{dat}{is the RD.experiment}

\item{levs}{this is the window region to update}

\item{cell}{are subsets of cell to incorporate}
}
\description{
RDVIew but for traces, also identifies the score of the neurons
}
