% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ReadDataDump.lee.2}
\alias{ReadDataDump.lee.2}
\title{Old school way to import data}
\usage{
ReadDataDump.lee.2(
  rd.name = NULL,
  img1 = NULL,
  img2 = NULL,
  img3 = NULL,
  img4 = NULL,
  img5 = NULL,
  img6 = NULL,
  img7 = NULL,
  img8 = NULL,
  fancy = F,
  fname = "Data (full).txt",
  wrdef = "wr1.docx",
  Wr = NULL,
  c.dat = "ROI Data.txt",
  sep = "\\t"
)
}
\description{
First step of this function is to change the working directory \code{setwd("Z:/Lee Leavitt/experimentFolder")}
Next use \code{ReadDataDump.lee.2("RD.date.age.gender.microscope.protocol", "bf.gfp.cy5")}
To make this work, you already need to have a \code{'Data (full).txt', 'ROI Data.txt', 'wr1.docx'}
}
