% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{RegionDeleter}
\alias{RegionDeleter}
\title{Something happened within you experiment and you need to delete a region of row fro the traces
to recover the data and clean up the display.
dat: This is the RD object
cell: Cell the display on the plot, if left empty, X.1 will be used
complete: logical, If you say true, all window regions will be reassessed. If False window region can be selected.}
\usage{
RegionDeleter(dat, cell = NULL, complete = T)
}
\description{
Something happened within you experiment and you need to delete a region of row fro the traces
to recover the data and clean up the display.
dat: This is the RD object
cell: Cell the display on the plot, if left empty, X.1 will be used
complete: logical, If you say true, all window regions will be reassessed. If False window region can be selected.
}
