% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{renamer}
\alias{renamer}
\title{Function to rename experiment to the name of the folder it
resides}
\usage{
renamer()
}
\description{
Function to rename experiment to the name of the folder it
resides
}
