% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roiReview.R
\name{ROIreview}
\alias{ROIreview}
\title{three tests Drop (confirm), Red (confirm) and Green (confirm)
return and RD object with the changes made to c.dat and bin}
\usage{
ROIreview(
  tmp,
  x.names = NULL,
  pad = 2,
  wh = 7,
  hh = 7,
  subset.n = 500,
  roi.img = NULL
)
}
\arguments{
\item{tmp}{is an RD object with images, "tritc.mean" and "gfp.mean" in c.dat}

\item{x.names}{is a list of specific cells to review}

\item{pad}{is the expansion factor about the center of the cell.}

\item{subset.n}{is number of cells to review at once instead of all at once.}
}
\description{
three tests Drop (confirm), Red (confirm) and Green (confirm)
return and RD object with the changes made to c.dat and bin
}
