% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{saveRD}
\alias{saveRD}
\title{Funciton to save the work along with create a unique savehistory}
\usage{
saveRD(dat)
}
\description{
Funciton to save the work along with create a unique savehistory
}
