% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{ScoreConstPharm.2}
\alias{ScoreConstPharm.2}
\title{calculate a table of cell characteristics globally and 
within specific windows
these specifics should include
mean and sd, sum of in window peaks, sum of out of window peaks}
\usage{
ScoreConstPharm.2(
  dat,
  t.type = NULL,
  snr = NULL,
  der = NULL,
  snr.lim = 3,
  blc.lim = 0.03,
  shws = 2
)
}
\description{
calculate a table of cell characteristics globally and 
within specific windows
these specifics should include
mean and sd, sum of in window peaks, sum of out of window peaks
}
