% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{ScoreConstPharm}
\alias{ScoreConstPharm}
\title{calculate a table of cell characteristics globally and 
within specific windows
these specifics should include
mean and sd, sum of in window peaks, sum of out of window peaks
1)  some measure of dead cell
2)  yes/no peak response for each window
3) peak height
4) max peak SNR
5) peak timing in window
6)
variance of smoothed - raw in window
define and number blank windows.}
\usage{
ScoreConstPharm(
  dat,
  blc = NULL,
  snr = NULL,
  der = NULL,
  snr.lim = 3,
  blc.lim = 0.03,
  shws = 2
)
}
\description{
calculate a table of cell characteristics globally and 
within specific windows
these specifics should include
mean and sd, sum of in window peaks, sum of out of window peaks
1)  some measure of dead cell
2)  yes/no peak response for each window
3) peak height
4) max peak SNR
5) peak timing in window
6)
variance of smoothed - raw in window
define and number blank windows.
}
