% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{TableBrewer}
\alias{TableBrewer}
\title{function to build a table with defined cell types, and selected collumns}
\usage{
TableBrewer(dat, ct.names = NULL, save = T, xlsx = T)
}
\description{
function to build a table with defined cell types, and selected collumns
}
