% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcd.R
\name{tcd}
\alias{tcd}
\title{TraceClick.d
ev(tcd)}
\usage{
tcd(
  dat,
  cells = NULL,
  img = dat$img1,
  l.img = c("img1"),
  yvar = FALSE,
  t.type = "t.dat",
  plot.new = F,
  info = T,
  pts = T,
  lns = T,
  bcex = 1,
  levs = NULL,
  klevs = NULL,
  sft = NULL,
  underline = T,
  zf = 20,
  lw = 2,
  sf = 1,
  dat.name = NULL,
  view_func_description = F,
  save_question = T
)
}
\arguments{
\item{up}{arrow move through list specified in entry}

\item{down}{arrow Move down through list spcified in entry}

\item{a}{place cell into the correct group in cell_types}

\item{1-0}{add cells to g.names1 through g.name10}

\item{q}{Quits the program}

\item{c}{add cells to g.names}

\item{d}{details for peakfunc}

\item{D}{LinesEvery seperation}

\item{f}{New trace fitting for pottassium pulses}

\item{F}{New smoothing factor for fit trace}

\item{ctrl-F}{function to correct the scoring}

\item{i}{Select image to display on Stacked Traces}

\item{I}{image for Multiview}

\item{l}{choose window region to display on stack trace plot}

\item{o}{order all cells in a new way}

\item{O}{order cells in Stacked Traces and multiview}

\item{p}{Toggles points on graph}

\item{P}{Pick a group/cells to click through}

\item{R}{reset group specified}

\item{r}{rename group names}

\item{s}{stack selected Groups}

\item{t}{brings up list of RD file. Select Trace (anything starting with t or mp)}

\item{u}{Underlines the Trace}

\item{v}{Show where cells are located and give zoomed in view}

\item{V}{choose cell info to display on traces}

\item{w}{Change Line Width on plot}

\item{x}{score this cell as a drop}

\item{X}{score cell as not dropped.}

\item{y}{Zoom yaxis automatically}

\item{z}{image zoom}

\item{F1}{advanced stat comparison statistic creator}

\item{F2}{advanced stat peak comparison min max norm creator}

\item{F3}{Density plot visualization}

\item{F5}{does something}
}
\description{
This function is an interactive dashboard for viewing your exeriments.
}
\details{
The output of this function returns a list of vectors of cell names
There are a few ways to input cell names into this program;

1)Character; ex. cells=c('X.1','X.2','X.3','X.4')
2)Numeric; ex. cells=c(1,2,3,4)
3)Character Lists; ex. active.cells[[1]]

Character lists/Cell groups, can be handled and displayed in a variety 
of ways. Using Keyboard commands (CASE SENSITVE);
}
