% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{testPulseFinder}
\alias{testPulseFinder}
\title{This is a function that returns a regular expression to find the testPulse}
\usage{
testPulseFinder(dat)
}
\description{
This is a function that returns a regular expression to find the testPulse
}
