% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePrep.R
\name{traceSticher}
\alias{traceSticher}
\title{This fucntion stiches two videos together}
\usage{
traceSticher(dat1Loc, dat2Loc, timeBuffer = 3, newName = NULL)
}
\arguments{
\item{dat1Loc}{is the flocation of the frist RD fil ex. "./1 video that yeeted itself/RD.1.Rdata"}

\item{dat2Loc}{is the second rd file taht needs to stich to "./2 R3J, TTAA, agonist vid/RD.2.Rdata"}

\item{timeBuffer}{is the time separation in minutes between the frist trace and the second trace}

\item{newName}{is the name of the experiment to save}
}
\description{
Make sure both Videos are in seperate Folders
ALSO HAVE SEPERATE WR1's that match  exactly what happened
1 Now tht we know the videos are ok we will simply do the cell profiler pipeline
2 we copy everything from the cell profiler pipeline into the 2 folders that contain the videos
3 Now do the pharming harvest and select both folders that contain the seperate videos
}
