% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{uncertaintyMaker}
\alias{uncertaintyMaker}
\title{This function calculates uncertainty for each probability data frame in the 
RD.experiment. Use probMaker and imageProbMaker prior to this or you will
get an error.}
\usage{
uncertaintyMaker(dat)
}
\description{
This function calculates uncertainty for each probability data frame in the 
RD.experiment. Use probMaker and imageProbMaker prior to this or you will
get an error.
}
