% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{WindowRepair_docx}
\alias{WindowRepair_docx}
\title{Made a mistake in you window regions?
If it is only the time sequence, but all other info is corrected
then make complete F.  This will allow you to select the windows that need reapri
if the naming is off, then make complete=F.  You will need to do a complete reapri
you will lose all information from RDView}
\usage{
WindowRepair_docx(dat, complete = F, trace_brew = F)
}
\description{
Made a mistake in you window regions?
If it is only the time sequence, but all other info is corrected
then make complete F.  This will allow you to select the windows that need reapri
if the naming is off, then make complete=F.  You will need to do a complete reapri
you will lose all information from RDView
}
