% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{XYtrace.2}
\alias{XYtrace.2}
\title{More advanced image clickers}
\usage{
XYtrace.2(
  dat,
  cells = NULL,
  img = NULL,
  cols = NULL,
  zoom = T,
  labs = T,
  yvar = F,
  zf = 40,
  t.type = NULL,
  sf = 1,
  plot.labs = T
)
}
\description{
More advanced image clickers
}
