% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{XYtrace}
\alias{XYtrace}
\title{Function allows for selection and deselection of cells to build stacked traces}
\usage{
XYtrace(dat, cell = NULL, img = NULL, cols = NULL, labs = F, y.var = T)
}
\description{
Function allows for selection and deselection of cells to build stacked traces
}
