% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{barPlotter}
\alias{barPlotter}
\title{Display census in a barplot}
\usage{
barPlotter(dat = NULL, cols = "Dark2", selectCT = T, horiz = T)
}
\arguments{
\item{dat}{this can either be and experiment RD.experiment or a vector of file paths see example}

\item{cols}{this is the colo.brewer pallette to use. \href{https://www.datanovia.com/en/wp-content/uploads/dn-tutorials/ggplot2/figures/0101-rcolorbrewer-palette-rcolorbrewer-palettes-colorblind-friendly-1.png}{click here}}

\item{selectCT}{logical T or F for cleaning up cell types}

\item{horiz}{logical T or F for how to plot the barplot}
}
\description{
Display census in a barplot
}
\examples{
\dontrun{
# First step make single csv files from each experiment. Follow this general pattern
RD.experiment <- census_to_table(RD.experiment)
# Only select a single thing. For example only select amp
TableBrewer(RD.experiment)
#
#
# Once it is saved, open the cvs and manually rename the amp/first row something like "CNF EP1 1uM"
# additionally you have full control over the data in the cvs files. I warn against changing any numbers
# unless you are summing together other tables.
#
# Once it is saved, open the cvs and manually rename the amp something like "CNF EP1 1uM"
#
# Set the working directory to have the correct location. This will be where the files are located
# In the example you can see that i've set the working directory to 
setwd('Y:/Cris Urcino/CNF experiments/Ep1/')
# This means i can access each csv file in the following way
tables <- c(
    "./190823.M.40.m3.p1 CNF-Ep1/amp.csv",
    "./200209.M.31.m3.p1 3uM.Ep1/amp.csv",
    "./200209.M.31.m3.p2 3uM.Ep1/amp.csv"
)
#
# Now barPlotter() it,
barPlotter(tables)
}
}
