% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceStats.R
\name{ideStatMaker}
\alias{ideStatMaker}
\title{Function taht creates a ide stat automatically if you have all control and test window regions in the data.
To make this work you can't half ass your window regions you must include all control window regions}
\usage{
ideStatMaker(dat, statType = "max", testPulseNames = "^[kK][.]30")
}
\arguments{
\item{statType}{this is the scp stat you would use}

\item{testPulseNames}{this is the test pulse names that contains the experiments}
}
\description{
Function taht creates a ide stat automatically if you have all control and test window regions in the data.
To make this work you can't half ass your window regions you must include all control window regions
}
