% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDView.R
\name{RDView}
\alias{RDView}
\title{tmp is an RD object
wr.i is the window region definition.
rd.name is the name of the RD object (used for png.out)
c.i is an alternate list of cells to review (instead of all)
rscale is a boolean for rescaling the data
wh is the window height
hh is the window width (why the hell did I name it hh?)}
\usage{
RDView(
  tmp,
  c.i = NULL,
  wr.i = "wr1",
  rd.name = NULL,
  rscale = F,
  wh = 11,
  hh = 6
)
}
\description{
tmp is an RD object
wr.i is the window region definition.
rd.name is the name of the RD object (used for png.out)
c.i is an alternate list of cells to review (instead of all)
rscale is a boolean for rescaling the data
wh is the window height
hh is the window width (why the hell did I name it hh?)
}
