% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{saveRD}
\alias{saveRD}
\title{Funciton to save the work along with create a unique savehistory
This function can now take a tmp.rd tmpRD or tmp as input. From there the
function looks for a RD.experiemnt in the global environemnt. If it finds 
one, it will save the RD.experiment as RD.experiment not tmp
If more than one RD.exepriment are in the work space it will stop and not save}
\usage{
saveRD(dat)
}
\description{
Funciton to save the work along with create a unique savehistory
This function can now take a tmp.rd tmpRD or tmp as input. From there the
function looks for a RD.experiemnt in the global environemnt. If it finds 
one, it will save the RD.experiment as RD.experiment not tmp
If more than one RD.exepriment are in the work space it will stop and not save
}
