% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTools.R
\name{traceProbMaker}
\alias{traceProbMaker}
\title{This function creates the binary scores for AITC, menthol, capsaicin and K40}
\usage{
traceProbMaker(dat)
}
\description{
This function creates the binary scores for AITC, menthol, capsaicin and K40
}
